+++ .github/workflows/execution.sh
@@ -5,6 +5,7 @@ EXECUTION_FILES=(
 #    ui_testing/testcases/extended_tests/test002_article.py
 
     ui_testing/testcases/basic_tests/test004_testunits.py
+    ui_testing/testcases/extended_tests/test003_testunits.py
 #    ui_testing/testcases/basic_tests/test003_testplans.py
 #    ui_testing/testcases/basic_tests/test005_contacts.py
 #    ui_testing/testcases/basic_tests/test006_orders.py
